backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
sqlalchemy==2.0.18
pydantic==2.0.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.11.1
pydantic-settings==2.0.1
email-validator==2.0.0
python-dotenv==1.0.0
